[HPSJAVA-636] Use standard SRS datacat Java client plus other minor changes and reorg.

+++ run-database/src/main/java/org/hps/run/database/RunSummary.java
@@ -97,7 +97,7 @@ public interface RunSummary {
      *
      * @return the total number of events in the run
      */
-    Integer getTotalEvents();
+    Long getTotalEvents();
 
     /**
      * Get the total number of EVIO files in this run.
